requirements.txt
pandas
numpy
matplotlib
Pillow==9.2.0
streamlit
tensorflow